Change all amount fields from float to int (amounts in cents)
- Change all amount fields from ?float to ?int in Amount classes
- Update SaleAmount, AuthAmount, RefundAmount, PostAuthAmount, BatchTotalAmount
- Update TipAdjustRequest tipAmount field
- Add comments explaining amount unit (cents)
- Update test cases to use integer amounts
- Update README documentation with amount unit examples

+++ src/Sunmi/Sunbay/Nexus/Model/Common/Amount.php
@@ -6,6 +6,7 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
 
 /**
  * Amount information
+ * All amount fields are in cents (e.g., 10000 = $100.00)
  *
  * @author Andy Li
  * @since 2025-12-19
@@ -13,12 +14,12 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
 class Amount
 {
     private ?string $priceCurrency = null;
-    private ?float $transAmount = null;
-    private ?float $orderAmount = null;
-    private ?float $taxAmount = null;
-    private ?float $surchargeAmount = null;
-    private ?float $tipAmount = null;
-    private ?float $cashbackAmount = null;
+    private ?int $transAmount = null;
+    private ?int $orderAmount = null;
+    private ?int $taxAmount = null;
+    private ?int $surchargeAmount = null;
+    private ?int $tipAmount = null;
+    private ?int $cashbackAmount = null;
     private ?string $pricingCurrency = null;
 
     public function getPriceCurrency(): ?string
@@ -32,67 +33,67 @@ class Amount
         return $this;
     }
 
-    public function getTransAmount(): ?float
+    public function getTransAmount(): ?int
     {
         return $this->transAmount;
     }
 
-    public function setTransAmount(?float $transAmount): self
+    public function setTransAmount(?int $transAmount): self
     {
         $this->transAmount = $transAmount;
         return $this;
     }
 
-    public function getOrderAmount(): ?float
+    public function getOrderAmount(): ?int
     {
         return $this->orderAmount;
     }
 
-    public function setOrderAmount(?float $orderAmount): self
+    public function setOrderAmount(?int $orderAmount): self
     {
         $this->orderAmount = $orderAmount;
         return $this;
     }
 
-    public function getTaxAmount(): ?float
+    public function getTaxAmount(): ?int
     {
         return $this->taxAmount;
     }
 
-    public function setTaxAmount(?float $taxAmount): self
+    public function setTaxAmount(?int $taxAmount): self
     {
         $this->taxAmount = $taxAmount;
         return $this;
     }
 
-    public function getSurchargeAmount(): ?float
+    public function getSurchargeAmount(): ?int
     {
         return $this->surchargeAmount;
     }
 
-    public function setSurchargeAmount(?float $surchargeAmount): self
+    public function setSurchargeAmount(?int $surchargeAmount): self
     {
         $this->surchargeAmount = $surchargeAmount;
         return $this;
     }
 
-    public function getTipAmount(): ?float
+    public function getTipAmount(): ?int
     {
         return $this->tipAmount;
     }
 
-    public function setTipAmount(?float $tipAmount): self
+    public function setTipAmount(?int $tipAmount): self
     {
         $this->tipAmount = $tipAmount;
         return $this;
     }
 
-    public function getCashbackAmount(): ?float
+    public function getCashbackAmount(): ?int
     {
         return $this->cashbackAmount;
     }
 
-    public function setCashbackAmount(?float $cashbackAmount): self
+    public function setCashbackAmount(?int $cashbackAmount): self
     {
         $this->cashbackAmount = $cashbackAmount;
         return $this;
