+++ src/Sunmi/Sunbay/Nexus/Model/Common/PostAuthAmount.php
@@ -8,57 +8,58 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
  * Post authorization amount information
  * Supports: orderAmount, tipAmount, taxAmount, surchargeAmount
  * Does NOT support: cashbackAmount
+ * All amount fields are in cents (e.g., 10000 = $100.00)
  *
  * @author Andy Li
  * @since 2025-12-19
  */
 class PostAuthAmount
 {
-    private ?float $orderAmount = null;
-    private ?float $tipAmount = null;
-    private ?float $taxAmount = null;
-    private ?float $surchargeAmount = null;
+    private ?int $orderAmount = null;
+    private ?int $tipAmount = null;
+    private ?int $taxAmount = null;
+    private ?int $surchargeAmount = null;
     private ?string $pricingCurrency = null;
 
-    public function getOrderAmount(): ?float
+    public function getOrderAmount(): ?int
     {
         return $this->orderAmount;
     }
 
-    public function setOrderAmount(?float $orderAmount): self
+    public function setOrderAmount(?int $orderAmount): self
     {
         $this->orderAmount = $orderAmount;
         return $this;
     }
 
-    public function getTipAmount(): ?float
+    public function getTipAmount(): ?int
     {
         return $this->tipAmount;
     }
 
-    public function setTipAmount(?float $tipAmount): self
+    public function setTipAmount(?int $tipAmount): self
     {
         $this->tipAmount = $tipAmount;
         return $this;
     }
 
-    public function getTaxAmount(): ?float
+    public function getTaxAmount(): ?int
     {
         return $this->taxAmount;
     }
 
-    public function setTaxAmount(?float $taxAmount): self
+    public function setTaxAmount(?int $taxAmount): self
     {
         $this->taxAmount = $taxAmount;
         return $this;
     }
 
-    public function getSurchargeAmount(): ?float
+    public function getSurchargeAmount(): ?int
     {
         return $this->surchargeAmount;
     }
 
-    public function setSurchargeAmount(?float $surchargeAmount): self
+    public function setSurchargeAmount(?int $surchargeAmount): self
     {
         $this->surchargeAmount = $surchargeAmount;
         return $this;
@@ -90,25 +91,25 @@ class PostAuthAmountBuilder
         $this->postAuthAmount = new PostAuthAmount();
     }
 
-    public function orderAmount(?float $orderAmount): self
+    public function orderAmount(?int $orderAmount): self
     {
         $this->postAuthAmount->setOrderAmount($orderAmount);
         return $this;
     }
 
-    public function tipAmount(?float $tipAmount): self
+    public function tipAmount(?int $tipAmount): self
     {
         $this->postAuthAmount->setTipAmount($tipAmount);
         return $this;
     }
 
-    public function taxAmount(?float $taxAmount): self
+    public function taxAmount(?int $taxAmount): self
     {
         $this->postAuthAmount->setTaxAmount($taxAmount);
         return $this;
     }
 
-    public function surchargeAmount(?float $surchargeAmount): self
+    public function surchargeAmount(?int $surchargeAmount): self
     {
         $this->postAuthAmount->setSurchargeAmount($surchargeAmount);
         return $this;
