+++ src/Sunmi/Sunbay/Nexus/Model/Common/RefundAmount.php
@@ -7,69 +7,70 @@ namespace Sunmi\Sunbay\Nexus\Model\Common;
 /**
  * Refund amount information
  * Supports: orderAmount, tipAmount, taxAmount, surchargeAmount, cashbackAmount
+ * All amount fields are in cents (e.g., 10000 = $100.00)
  *
  * @author Andy Li
  * @since 2025-12-19
  */
 class RefundAmount
 {
-    private ?float $orderAmount = null;
-    private ?float $tipAmount = null;
-    private ?float $taxAmount = null;
-    private ?float $surchargeAmount = null;
-    private ?float $cashbackAmount = null;
+    private ?int $orderAmount = null;
+    private ?int $tipAmount = null;
+    private ?int $taxAmount = null;
+    private ?int $surchargeAmount = null;
+    private ?int $cashbackAmount = null;
     private ?string $pricingCurrency = null;
 
-    public function getOrderAmount(): ?float
+    public function getOrderAmount(): ?int
     {
         return $this->orderAmount;
     }
 
-    public function setOrderAmount(?float $orderAmount): self
+    public function setOrderAmount(?int $orderAmount): self
     {
         $this->orderAmount = $orderAmount;
         return $this;
     }
 
-    public function getTipAmount(): ?float
+    public function getTipAmount(): ?int
     {
         return $this->tipAmount;
     }
 
-    public function setTipAmount(?float $tipAmount): self
+    public function setTipAmount(?int $tipAmount): self
     {
         $this->tipAmount = $tipAmount;
         return $this;
     }
 
-    public function getTaxAmount(): ?float
+    public function getTaxAmount(): ?int
     {
         return $this->taxAmount;
     }
 
-    public function setTaxAmount(?float $taxAmount): self
+    public function setTaxAmount(?int $taxAmount): self
     {
         $this->taxAmount = $taxAmount;
         return $this;
     }
 
-    public function getSurchargeAmount(): ?float
+    public function getSurchargeAmount(): ?int
     {
         return $this->surchargeAmount;
     }
 
-    public function setSurchargeAmount(?float $surchargeAmount): self
+    public function setSurchargeAmount(?int $surchargeAmount): self
     {
         $this->surchargeAmount = $surchargeAmount;
         return $this;
     }
 
-    public function getCashbackAmount(): ?float
+    public function getCashbackAmount(): ?int
     {
         return $this->cashbackAmount;
     }
 
-    public function setCashbackAmount(?float $cashbackAmount): self
+    public function setCashbackAmount(?int $cashbackAmount): self
     {
         $this->cashbackAmount = $cashbackAmount;
         return $this;
@@ -101,31 +102,31 @@ class RefundAmountBuilder
         $this->refundAmount = new RefundAmount();
     }
 
-    public function orderAmount(?float $orderAmount): self
+    public function orderAmount(?int $orderAmount): self
     {
         $this->refundAmount->setOrderAmount($orderAmount);
         return $this;
     }
 
-    public function tipAmount(?float $tipAmount): self
+    public function tipAmount(?int $tipAmount): self
     {
         $this->refundAmount->setTipAmount($tipAmount);
         return $this;
     }
 
-    public function taxAmount(?float $taxAmount): self
+    public function taxAmount(?int $taxAmount): self
     {
         $this->refundAmount->setTaxAmount($taxAmount);
         return $this;
     }
 
-    public function surchargeAmount(?float $surchargeAmount): self
+    public function surchargeAmount(?int $surchargeAmount): self
     {
         $this->refundAmount->setSurchargeAmount($surchargeAmount);
         return $this;
     }
 
-    public function cashbackAmount(?float $cashbackAmount): self
+    public function cashbackAmount(?int $cashbackAmount): self
     {
         $this->refundAmount->setCashbackAmount($cashbackAmount);
         return $this;
